blockchain_integratio/transcendence_root/app/pingpong/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
channels
daphne
channels-redis
# blockchain
web3==6.15.1